refactor voice search endpoint
