+++ core_backend/app/question_answer/speech_components/utils.py
@@ -1,9 +1,10 @@
 import os
 
+from fastapi import HTTPException
 from pydub import AudioSegment
 
 from ...llm_call.llm_prompts import IdentifiedLanguage
-from ...utils import setup_logger
+from ...utils import get_http_client, setup_logger
 
 logger = setup_logger("Voice utils")
 
@@ -80,3 +81,16 @@ def set_wav_specifications(wav_filename: str) -> str:
 
     logger.info(f"Updated file created: {updated_wav_filename}")
     return updated_wav_filename
+
+
+async def post_to_speech(file_path: str, endpoint_url: str) -> dict:
+    """
+    Post request the file to the speech endpoint to get the transcription
+    """
+    async with get_http_client() as client:
+        async with client.post(endpoint_url, json={"file_path": file_path}) as response:
+            if response.status != 200:
+                error_content = await response.json()
+                logger.error(f"Error from CUSTOM_SPEECH_ENDPOINT: {error_content}")
+                raise HTTPException(status_code=response.status, detail=error_content)
+            return await response.json()
